Don't delete admin action webauthn challenges on first consumption.

+++ lib/auth/webauthn/login.go
@@ -319,10 +319,15 @@ func (f *loginFlow) finish(ctx context.Context, user string, resp *wantypes.Cred
 		return nil, "", trace.Wrap(err)
 	}
 
-	// The user just solved the challenge, so let's make sure it won't be used
-	// again.
-	if err := f.sessionData.Delete(ctx, user, challenge); err != nil {
-		log.Warnf("WebAuthn: failed to delete login SessionData for user %v (passwordless = %v)", user, passwordless)
+	// The user just solved the challenge, so let's make sure it won't be used again.
+	if f.Scope == wanpb.Scope_SCOPE_ADMIN_ACTION {
+		// For admin action MFA responses, we don't delete the response immediately. Instead
+		// add a new expiration time so it can be reused for a short duration.
+		f.sessionData.Upsert(ctx, user, sessionDataPB)
+	} else {
+		if err := f.sessionData.Delete(ctx, user, challenge); err != nil {
+			log.Warnf("WebAuthn: failed to delete login SessionData for user %v (passwordless = %v)", user, passwordless)
+		}
 	}
 
 	return dev, user, nil
